LPS-169073 Move the usage of abstract method createPortalCache(PortalCacheConfiguration, boolean) in BasePortalCacheManager to EhcachePortalCacheManager to remove the abstract method from BasePortalCacheManager
