+++ modules/apps/portal-cache/portal-cache-ehcache-impl/src/main/java/com/liferay/portal/cache/ehcache/internal/BasePortalCacheManager.java
@@ -14,15 +14,10 @@
 
 package com.liferay.portal.cache.ehcache.internal;
 
-import com.liferay.petra.string.CharPool;
-import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.cache.AggregatedPortalCacheManagerListener;
-import com.liferay.portal.cache.LowLevelCache;
-import com.liferay.portal.cache.MVCCPortalCache;
 import com.liferay.portal.cache.PortalCacheListenerFactory;
 import com.liferay.portal.cache.PortalCacheManagerListenerFactory;
-import com.liferay.portal.cache.TransactionalPortalCache;
 import com.liferay.portal.cache.configuration.PortalCacheConfiguration;
 import com.liferay.portal.cache.configuration.PortalCacheManagerConfiguration;
 import com.liferay.portal.kernel.cache.PortalCache;
@@ -31,8 +26,6 @@ import com.liferay.portal.kernel.cache.PortalCacheListener;
 import com.liferay.portal.kernel.cache.PortalCacheListenerScope;
 import com.liferay.portal.kernel.cache.PortalCacheManager;
 import com.liferay.portal.kernel.cache.PortalCacheManagerListener;
-import com.liferay.portal.kernel.model.MVCCModel;
-import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 
 import java.io.Serializable;
@@ -76,44 +69,6 @@ public abstract class BasePortalCacheManager<K extends Serializable, V>
 	}
 
 	@Override
-	public PortalCache<K, V> getPortalCache(
-			String portalCacheName, boolean mvcc, boolean sharded)
-		throws PortalCacheException {
-
-		return portalCaches.compute(
-			portalCacheName,
-			(key, value) -> {
-				if (value != null) {
-					_verifyMVCCPortalCache(value, mvcc);
-					_verifyShardedPortalCache(value, sharded);
-
-					return value;
-				}
-
-				PortalCacheConfiguration portalCacheConfiguration =
-					portalCacheManagerConfiguration.getPortalCacheConfiguration(
-						portalCacheName);
-
-				value = createPortalCache(portalCacheConfiguration, sharded);
-
-				initPortalCacheListeners(value, portalCacheConfiguration);
-
-				if (mvcc) {
-					value = (PortalCache<K, V>)new MVCCPortalCache<>(
-						(LowLevelCache<K, MVCCModel>)value);
-				}
-
-				if (transactionalPortalCacheEnabled &&
-					_isTransactionalPortalCache(portalCacheName)) {
-
-					value = new TransactionalPortalCache<>(value, mvcc);
-				}
-
-				return value;
-			});
-	}
-
-	@Override
 	public Set<PortalCacheManagerListener> getPortalCacheManagerListeners() {
 		return aggregatedPortalCacheManagerListener.
 			getPortalCacheManagerListeners();
@@ -153,9 +108,6 @@ public abstract class BasePortalCacheManager<K extends Serializable, V>
 	public void unregisterPortalCacheManagerListeners() {
 		aggregatedPortalCacheManagerListener.clearAll();
 	}
-
-	protected abstract PortalCache<K, V> createPortalCache(
-		PortalCacheConfiguration portalCacheConfiguration, boolean sharded);
 
 	protected abstract void doDestroy();
 
@@ -238,90 +190,5 @@ public abstract class BasePortalCacheManager<K extends Serializable, V>
 		new ConcurrentHashMap<>();
 	protected boolean transactionalPortalCacheEnabled;
 	protected String[] transactionalPortalCacheNames = StringPool.EMPTY_ARRAY;
-
-	private boolean _isTransactionalPortalCache(String portalCacheName) {
-		for (String namePattern : transactionalPortalCacheNames) {
-			if (StringUtil.wildcardMatches(
-					portalCacheName, namePattern, CharPool.QUESTION,
-					CharPool.STAR, CharPool.PERCENT, true)) {
-
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	private void _verifyMVCCPortalCache(
-		PortalCache<K, V> portalCache, boolean mvcc) {
-
-		if (mvcc == portalCache.isMVCC()) {
-			return;
-		}
-
-		StringBundler sb = new StringBundler(9);
-
-		sb.append("Unable to get portal cache ");
-		sb.append(portalCache.getPortalCacheName());
-		sb.append(" from portal cache manager ");
-		sb.append(portalCacheManagerName);
-		sb.append(" as a ");
-
-		if (mvcc) {
-			sb.append("MVCC ");
-		}
-		else {
-			sb.append("non-MVCC ");
-		}
-
-		sb.append("portal cache, cause a ");
-
-		if (portalCache.isMVCC()) {
-			sb.append("MVCC ");
-		}
-		else {
-			sb.append("non-MVCC ");
-		}
-
-		sb.append("portal cache with same name exists.");
-
-		throw new IllegalStateException(sb.toString());
-	}
-
-	private void _verifyShardedPortalCache(
-		PortalCache<K, V> portalCache, boolean sharded) {
-
-		if (sharded == portalCache.isSharded()) {
-			return;
-		}
-
-		StringBundler sb = new StringBundler(9);
-
-		sb.append("Unable to get portal cache ");
-		sb.append(portalCache.getPortalCacheName());
-		sb.append(" from portal cache manager ");
-		sb.append(portalCacheManagerName);
-		sb.append(" as a ");
-
-		if (sharded) {
-			sb.append("sharded ");
-		}
-		else {
-			sb.append("nonsharded ");
-		}
-
-		sb.append("portal cache, cause a ");
-
-		if (portalCache.isSharded()) {
-			sb.append("sharded ");
-		}
-		else {
-			sb.append("nonsharded ");
-		}
-
-		sb.append("portal cache with same name exists.");
-
-		throw new IllegalStateException(sb.toString());
-	}
 
 }
